backend/requirements.txt
channels[daphne]==4.2.*
channels_redis==4.2.*
django==5.1.*
django-colorfield==0.11.*
django-cors-headers==4.6.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.4.*
environs==14.1.*
gunicorn==23.0.*
openpyxl==3.1.*
pillow==11.1.*
psycopg2-binary==2.9.*
